requirements.txt
Flask==3.0.3
pycryptodome==3.20.0
